pyproject.toml
[tool.poetry]
name = "fcontrol-api"
version = "0.1.0"
description = ""
authors = ["Raphael R Caldas <raphael008@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
fastapi = { extras = ["standard"], version = "0.*" }
pydantic = { extras = ["email"], version = "^2.8.2" }
sqlalchemy = { extras = ["mypy"], version = "^2.0.37" }
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
pwdlib = { extras = ["argon2"], version = "0.*" }
tzdata = "^2024.1"
python-multipart = "0.*"
taskipy = "^1.13"
psycopg = { extras = ["binary"], version = "^3.2.1" }
psycopg2-binary = "^2.9.10"
httptools = "^0.6.4"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
ruff = "0.*"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
httpx = "0.*"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
aiosqlite = "^0.20.0"
trio = "^0.28.0"
pytest-asyncio = "^0.25.2"
gevent = "^24.11.1"
tqdm = "^4.66.0"
testcontainers = {extras = ["postgres"], version = "^4.13.3"}

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR0913', 'PLR0915', 'PLR0914', 'PLR0917', 'PLR0912']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
run = 'fastapi run fcontrol_api/app.py --reload --host 0.0.0.0 --port 8000'

pre_test = 'task lint'
test = 'pytest -s -x --cov=fcontrol_api -vv'
post_test = 'coverage html'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
